Split monitoring to its own file
