+++ src/js/hds-cc/hds-cc.js
@@ -17,7 +17,7 @@ import {
   getTranslationKeys,
 } from './hds-cc_translations';
 
-import deleteCookie from './deleteCookie';
+import MonitorAncCleanBrowserStorages from './monitorAndCleanBrowserStorages';
 
 /**
  * Represents a class for managing cookie consent.
@@ -39,9 +39,7 @@ class HdsCookieConsentClass {
 
   #SETTINGS_PAGE_SELECTOR;
 
-  #MONITOR_INTERVAL;
-
-  #REMOVE;
+  #MONITOR;
 
   #TEMP_CSS_PATH;
 
@@ -66,33 +64,6 @@ class HdsCookieConsentClass {
   };
 
   #cookie_name = 'city-of-helsinki-cookie-consents'; // Overridable default value
-
-  // Initial keys found when script is initialized
-  #INITIAL_STORED_KEYS = {
-    cookie: [],
-    localStorage: [],
-    sessionStorage: [],
-    indexedDB: [],
-    cacheStorage: [],
-  };
-
-  // Keys already reported via event
-  #reportedKeys = {
-    cookie: [],
-    localStorage: [],
-    sessionStorage: [],
-    indexedDB: [],
-    cacheStorage: [],
-  };
-
-  // If key removal was not successfull, add it to this list to prevent multiple tries
-  #removalFailedKeys = {
-    cookie: [],
-    localStorage: [],
-    sessionStorage: [],
-    indexedDB: [],
-    cacheStorage: [],
-  };
 
   // MARK: Public methods
   /**
@@ -135,20 +106,14 @@ class HdsCookieConsentClass {
     this.#PAGE_CONTENT_SELECTOR = pageContentSelector;
     this.#SUBMIT_EVENT = submitEvent;
     this.#SETTINGS_PAGE_SELECTOR = settingsPageSelector;
-    this.#MONITOR_INTERVAL = monitorInterval;
-    this.#REMOVE = remove;
     this.#TEMP_CSS_PATH = tempCssPath;
-
-    this.#INITIAL_STORED_KEYS = {
-      cookie: this.#getCookieNamesArray(),
-      localStorage: Object.keys(localStorage),
-      sessionStorage: Object.keys(sessionStorage),
-      indexedDB: this.#getIndexedDBNamesArray(),
-      cacheStorage: this.#getCacheStorageNamesString(),
-    };
 
     window.hds = window.hds || {};
     window.hds.cookieConsent = this;
+
+    if (monitorInterval > 0) {
+      this.#MONITOR = new MonitorAncCleanBrowserStorages(monitorInterval, remove);
+    }
 
     if (document.readyState === 'loading') {
       document.addEventListener('DOMContentLoaded', () => {
@@ -822,261 +787,6 @@ class HdsCookieConsentClass {
     }
   }
 
-  // MARK: Monitoring
-
-  /**
-   * Returns a string containing the names of all cookies.
-   * @private
-   * @return {array} An array containing the names of all cookies.
-   */
-  #getCookieNamesArray() {
-    const cookies = document.cookie.split(';');
-    const cookieNames = cookies.map(cookie => cookie.split('=')[0].trim());
-    return cookieNames;
-  }
-
-
-  /**
-   * Retrieves the names of all indexedDB databases as a string.
-   * @private
-   * @return {Promise<array>} A promise that resolves to an array containing the names of all indexedDB databases. If there are no indexedDB databases, an empty array is returned.
-   */
-  async #getIndexedDBNamesArray() {
-    if (indexedDB && indexedDB.databases) {
-      const databases = await indexedDB.databases();
-      const databaseNames = databases.map(db => db.name);
-      return databaseNames;
-    }
-    return [];
-  }
-
-  /**
-   * Retrieves the names of all cache storages as a string.
-   * @private
-   * @return {Promise<array>} A promise that resolves to an array containing the names of all cache storages. If there are no cache storages, an empty array is returned.
-   */
-  async #getCacheStorageNamesString() {
-    if (caches) {
-      const cacheNames = await caches.keys();
-      return cacheNames;
-    }
-    return [];
-  }
-
-  /**
-   * Checks if a given key is consented based on the provided consented keys with wildcards.
-   *
-   * @param {string} key - The key to check consent for.
-   * @param {Array<string>} consentedKeys - The array of consented keys.
-   * @return {boolean} - Returns true if the key is consented, false otherwise.
-   */
-  #isKeyConsented(key, consentedKeys) {
-    // If no keys are consented, return false
-    if (!Array.isArray(consentedKeys) || consentedKeys.length === 0) {
-      return false;
-    }
-    // Check if the key is directly consented
-    if (consentedKeys.includes(key)) {
-      return true;
-    }
-
-    // Check if the key matches a wildcard pattern in consentedKeys that have * in them
-    const consentedKeysWithWildcard = consentedKeys.filter(consentedKey => consentedKey.includes('*'));
-    const consentedKeysRegexp = consentedKeysWithWildcard.map(consentedKey => new RegExp(`^${consentedKey.replace(/[.+?^${}()|[\]\\]/g, '\\$&').replace(/\*/g, '.*')}$`));
-
-    // Check if the key matches any of the wildcard patterns
-    return consentedKeysRegexp.some(regexp => regexp.test(key));
-  }
-
-  /**
-   * Monitors the consented keys and reports any unapproved keys. (cookies or localStorage or sessionStorage)
-   * Reports found unapproved keys to console and dispatches an event based on type.
-   * @private
-   * @param {string} typeString - The type of keys being monitored.
-   * @param {string[]} consentedKeysArray - An array of consented keys.
-   * @param {string[]} initialStoredKeysArray - The initial stored keys.
-   * @param {string[]} reportedKeysArray - An array of reported keys.
-   * @param {string[]} currentStoredKeysArray - An array of current stored keys.
-   * @param {string} consentedGroups - The consented groups.
-   */
-  #monitor(
-    typeString,
-    consentedKeysArray,
-    initialStoredKeysArray,
-    reportedKeysArray,
-    currentStoredKeysArray,
-    consentedGroups,
-  ) {
-    if (currentStoredKeysArray.join(';') !== initialStoredKeysArray.join(';')) {
-
-      // Find items that appear only in currentStoredKeysArray and filter out the ones that are already in consentedKeysArray
-      const unapprovedKeys = currentStoredKeysArray.filter((key) => {
-        if (
-          key === '' || // If the key is empty, filter it out
-          initialStoredKeysArray.includes(key) || // If key is not new, filter it out
-          reportedKeysArray.includes(key) || // If key is already reported, filter it out
-          this.#isKeyConsented(key, consentedKeysArray) // If key is consented (with possible wildcards), filter it out
-        ) {
-          return false;
-        }
-        return true;
-      });
-
-      if (unapprovedKeys.length > 0) {
-        console.log(`Cookie consent found unapproved ${typeString}(s): '${unapprovedKeys.join('\', \'')}'`);
-
-        const event = new CustomEvent('hds-cookie-consent-unapproved-item-found', {
-          detail: {
-            type: typeString,
-            keys: unapprovedKeys,
-            consentedGroups
-          }
-        });
-        window.dispatchEvent(event);
-
-        reportedKeysArray.push(...unapprovedKeys);
-      }
-    }
-
-    if (this.#REMOVE) {
-      const deleteKeys = currentStoredKeysArray.filter((key) => {
-        if (
-          key === '' || // If the key is empty, filter it out
-          this.#isKeyConsented(key, consentedKeysArray) // If key is consented (with possible wildcards), filter it out
-        ) {
-          return false;
-        }
-        return true;
-      });
-
-      if(deleteKeys.length > 0) {
-        // console.log('deleteKeys', deleteKeys, deleteKeys.length);
-        deleteKeys.forEach(key => {
-          // console.log('typeString', typeString, this.#removalFailedKeys, this.#removalFailedKeys[typeString]);
-          if (!this.#removalFailedKeys[typeString].includes(key)) {
-            console.log(`Cookie consent will delete unapproved ${typeString}(s): '${deleteKeys.join('\', \'')}'`);
-
-            if (typeString === 'cookie') {
-              deleteCookie(key);
-              if (this.#getCookie(key)) {
-                console.error(`Error deleting cookie '${key}' will ignore it for now`);
-                this.#removalFailedKeys.cookie.push(key);
-              }
-            } else if (typeString === 'localStorage') {
-              localStorage.removeItem(key);
-            } else if (typeString === 'sessionStorage') {
-              sessionStorage.removeItem(key);
-            } else if (typeString === 'indexedDB') {
-              const request = indexedDB.deleteDatabase(key);
-              request.onsuccess = () => {
-                // console.log(`IndexedDB database '${key}' deleted successfully.`);
-                // Remove the key from the blacklist as the deletion was successful
-                this.#removalFailedKeys.indexedDB = this.#removalFailedKeys.indexedDB.filter(item => item !== key);
-              };
-              request.onerror = () => {
-                // console.error(`Error deleting IndexedDB database '${key}'`);
-                this.#removalFailedKeys.indexedDB.push(key);
-              };
-              request.onblocked = () => {
-                // console.warn(`IndexedDB database '${key}' deletion blocked.`);
-                this.#removalFailedKeys.indexedDB.push(key);
-              };
-            } else if (typeString === 'cacheStorage') {
-              caches.delete(key).then((response) => {
-                if(response) {
-                  console.log(`Cache '${key}' has been deleted`);
-                } else {
-                  console.log(`Cache '${key}' not found`);
-                }
-              });
-            }
-          }
-        });
-      }
-      }
-
-  }
-
-  /**
-   * Monitors cookies, local storage, and session storage for unconsented new keys.
-   * @private
-   */
-  async #monitorLoop() {
-    // console.log('monitoring', JSON.stringify(this.#reportedKeys));
-
-    const consentCookie = this.#getCookie();
-    let acceptedGroups = '';
-    if (consentCookie && consentCookie.groups) {
-      acceptedGroups = Object.keys(consentCookie.groups).join(';');
-    }
-
-    const consentedCookies = consentCookie?.cookies?.split(';') || [];
-    consentedCookies.push(this.#cookie_name);
-    this.#monitor(
-      'cookie',
-      consentedCookies,
-      this.#INITIAL_STORED_KEYS.cookie,
-      this.#reportedKeys.cookie,
-      this.#getCookieNamesArray(),
-      acceptedGroups,
-    );
-
-    const consentedLocalStorage = consentCookie?.localStorage || [];
-    this.#monitor(
-      'localStorage',
-      consentedLocalStorage,
-      this.#INITIAL_STORED_KEYS.localStorage,
-      this.#reportedKeys.localStorage,
-      Object.keys(localStorage),
-      acceptedGroups,
-    );
-
-    const consentedSessionStorage = consentCookie?.sessionStorage || [];
-    this.#monitor(
-      'sessionStorage',
-      consentedSessionStorage,
-      this.#INITIAL_STORED_KEYS.sessionStorage,
-      this.#reportedKeys.sessionStorage,
-      Object.keys(sessionStorage),
-      acceptedGroups,
-    );
-
-    if (indexedDB) {
-      const consentedIndexedDB = consentCookie?.indexedDB || [];
-      this.#monitor(
-        'indexedDB',
-        consentedIndexedDB,
-        (await this.#INITIAL_STORED_KEYS.indexedDB),
-        this.#reportedKeys.indexedDB,
-        (await this.#getIndexedDBNamesArray()),
-        acceptedGroups,
-      );
-    }
-
-    if (caches) {
-      const consentedCacheStorage = consentCookie?.cacheStorage || [];
-      this.#monitor(
-        'cacheStorage',
-        consentedCacheStorage,
-        (await this.#INITIAL_STORED_KEYS.cacheStorage),
-        this.#reportedKeys.cacheStorage,
-        (await this.#getCacheStorageNamesString()),
-        acceptedGroups,
-      );
-    }
-  }
-
-  /**
-   * Monitors cookies and storage at a specified interval.
-   * @private
-   */
-  #monitorCookiesAndStorage() {
-    const interval = Math.max(this.#MONITOR_INTERVAL, 50);
-
-    this.#monitorLoop();
-    setInterval(() => { this.#monitorLoop(); }, interval);
-  }
-
   // MARK: Initializer
 
   /**
@@ -1109,8 +819,8 @@ class HdsCookieConsentClass {
       }
     }
 
-    if (this.#MONITOR_INTERVAL > 0) {
-      this.#monitorCookiesAndStorage();
+    if (this.#MONITOR) {
+      this.#MONITOR.init(this.#cookie_name);
     }
   };
 }
